Added "try: except KeyError:" to every block reading data that could be in the ODB or not
Mod 4) Added "try: except KeyError:" to every block reading data that could be in the ODB or not, so when the data is not there it just ignores that type of data. Beware that the script does not look for every conceivable type of data that could be on the ODB, so later we may have to add more code to handle that data. Future work: reduce VTK file size by writing only data available on ODB.

Mod 5) Added block to read "strains" (keyword "E").  

Mod 6) Added code to handle ParView type 9 elements (Quads)
                        if mesh_type==10 or mesh_type==12: #only continuum elements have 6 stress values
                            L1[valueX.nodeLabel-1][5] += valueX.data[4]
                            L1[valueX.nodeLabel-1][6] += valueX.data[5]
                        

Mod 7) Restored Function 2 to its original form. I don't want to update 2 identical functions. In the future I will try to merge both functions into one single function that can be called with or without parameters. 

Mod 8) Exhausted frames, just break the loop, no need to kill the Python shell/
                #Detect whether the input frame is out of range
                try:
                    TRY = odb.steps[stepname].frames[int(i_frame)]
                except:
                    print "* warning * input frame exceeds the range of frames" 
                    # os._exit(0) #just break the loop, no need to kill the Python shell
                    break