Synch'd common code for Functions 1 and 2
Synch'd common code for Functions 1 and 2. It is sad to have the same code repeated. Still trying to figure out a way to make just one function that does the job of both. Function1 is tested with piecenum=1, elements 9 and 12, several keywords missing in the .odb, either because they were not requested in Abaqus or because even if requested the type of analysis does not support them. If a keyword is requested and supported by the data is null, it will be copied to .vtk as null, wasting space. The solution is to remove the request from the FieldOutputRequest in Abaqus.